Fix bug in detecting player movement state
Two factors resulted in player movement state not being tracked
correctly:

1. CommonLibSSE was updated
2. Movement state was checked in ascending order of movement level

It is not clear why the behaviour of the movement state flags is
observed to be different on the newer version of CommonLibSSE.

Movement state needs to be checked in descending order because multiple
state flags may be active at once.

Reverting commit 7783773d8c0236e806ebad21251eb08fc73dcaf2 and reversing
the order in which movement state is checked resolves the issue.